Updated Non-Player Condition
Since we dont want players, we need to check that we dont have players in the array. 
Condition to check is "is _x a player?" and only act if the result is false. Therefore, setting the false option to trigger the removal of _x from the array as "!(isPlayer _x)" is the better way.